workspace/investment-game/services/controller/requirements.txt
flask
ulid-py